requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.9
dotenv==0.19.0
idna==3.4
mysql-connector-python==8.0.30
protobuf==3.20.1
requests==2.28.1
urllib3==1.26.12
